@@ -5,6 +5,7 @@
   "private": true,
   "dependencies": {
     "@fontsource/nanum-gothic": "^5.0.7",
+    "@tanstack/react-query": "^5.56.2",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
@@ -13,9 +14,9 @@
     "http-proxy-middleware": "^2.0.6",
     "moment": "^2.29.4",
     "moment-timezone": "^0.5.43",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-helmet-async": "^1.3.0",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
