@@ -1,20 +1,28 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 import { HelmetProvider } from "react-helmet-async";
 
-ReactDOM.render(
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+/** React Query Client */
+const queryClient = new QueryClient();
+
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
   <BrowserRouter basename="/namu-soup">
     <HelmetProvider>
-      <React.StrictMode>
-        <App />
-      </React.StrictMode>
+      <QueryClientProvider client={queryClient}>
+        <React.StrictMode>
+          <App />
+        </React.StrictMode>
+      </QueryClientProvider>
     </HelmetProvider>
-  </BrowserRouter>,
-  document.getElementById("root")
+  </BrowserRouter>
 );
 
 // If you want to start measuring performance in your app, pass a function
